requirements.txt
Flask==1.1.1
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Keras==2.2.4
numpy==1.16.5
pyflakes==2.1.1
PyMySQL==0.9.3
tensorflow==1.13.1
tqdm==4.36.1
